add more get requests, including the start details object

+++ Get-SureFlapToken.ps1
@@ -8,6 +8,10 @@ $postParams = @{
  "device_id" = "$device_id"
 } | ConvertTo-Json
 
-$res = Invoke-RestMethod -Method Post -Uri $uri -Body $postParams -ContentType "application/json"
-$token = $res.data.token
-$token | Out-File "sureflap.token"
+if (Test-Path .\sureflap.token) {
+    $token = Get-Content .\sureflap.token
+} else {
+    $res = Invoke-RestMethod -Method Post -Uri $uri -Body $postParams -ContentType "application/json"
+    $token = $res.data.token
+    $token | Out-File "sureflap.token"
+}
